Common code refactor
